Criação de metodos de criação de aluno, organização de camadas do projeto e criação da interface de repositorio.
