--- src/main/java/br/com/danichs/Indicacao.java
@@ -1,12 +1,0 @@
-package br.com.danichs;
-
-public class Indicacao {
-
-    private Aluno indicado;
-    private Aluno indicante;
-
-    public Indicacao(Aluno indicado, Aluno indicante){
-        this.indicado = indicado;
-        this.indicante = indicante;
-    }
-}
