+++ src/main/java/br/com/danichs/dominio/aluno/CPF.java
@@ -1,4 +1,4 @@
-package br.com.danichs;
+package br.com.danichs.dominio.aluno;
 
 //Value object, não possui um identificador para diferenciação de cada objeto, se há mesmo endereço por exemplo, significa ser o mesmo endereço.
 public class CPF {
