+++ src/main/java/br/com/danichs/dominio/aluno/RepositorioDeAlunos.java
@@ -1,0 +1,5 @@
+package br.com.danichs.dominio.aluno;
+
+public interface RepositorioDeAlunos {
+
+}
