+++ src/main/java/br/com/danichs/dominio/indicacao/Indicacao.java
@@ -1,0 +1,30 @@
+package br.com.danichs.dominio.indicacao;
+
+import java.time.LocalDateTime;
+
+import br.com.danichs.dominio.aluno.Aluno;
+
+public class Indicacao {
+
+    private Aluno indicado;
+    private Aluno indicante;
+    private LocalDateTime dataIndicacao;
+
+    public Indicacao(Aluno indicado, Aluno indicante){
+        this.indicado = indicado;
+        this.indicante = indicante;
+        this.dataIndicacao = LocalDateTime.now();
+    }
+
+    public Aluno getIndicado() {
+        return indicado;
+    }
+
+    public Aluno getIndicante() {
+        return indicante;
+    }
+
+    public LocalDateTime getDataIndicacao() {
+        return dataIndicacao;
+    }
+}
